added fitness test updates and week changes

+++ src/main/java/dcu/DAO/FitnessTestDAO.java
@@ -16,8 +16,8 @@ public interface FitnessTestDAO {
     public List<FitnessTest> getFintessTestById(final int fitnessId);
     public List<FitnessTest> getFintessTestByPlayerId(final int playerId);
     //public void createFitnessTest(final int playerId, final int bodyFat);
-    public void createFitnessTest(final int chinUps,int pushUps,final int playerId,final int sprint,final int kmRun,final int agility,final int workLifeStyle,final int weeklyAverageTrainingCount);
-    public void updateFitnessTest(final int chinUps,int pushUps,final int playerId,final int sprint,final int kmRun,final int agility,final int workLifestyle,final int weeklyAverageTrainingCount);
+    public void createFitnessTest(final int chinUps,int pushUps,final int playerId,final int sprint,final int kmRun,final int agility,final int workLifeStyle,final double weeklyAverageTrainingCount);
+    public void updateFitnessTest(final int chinUps,int pushUps,final int playerId,final int sprint,final int kmRun,final int agility,final int workLifestyle,final double weeklyAverageTrainingCount);
 
 
     }
